add client mod
